Code Improvements
Made a lot of code improvements
