+++ package.json
@@ -38,7 +38,6 @@
     "react-dom": "^18.2.0",
     "react-timeago": "^7.1.0",
     "seedrandom": "^3.0.5",
-    "sharp": "^0.32.0",
     "styled-components": "^5.1.1"
   },
   "devDependencies": {
