+++ src/components/common/layout/search-results.tsx
