+++ src/pages/404.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import NotFoundPage from "@client/components/Pages/Error/NotFound";
+import NotFoundPage from "@client/components/pages/error/not-found";
 import React from "react";
 
 const NotFound: FC = () => {
